Added L476RG

+++ bootutil/inc/com.h
@@ -2,7 +2,6 @@
 #define COM_H_
 
 #include <stdint.h>
-#include <stddef.h>
 
 #include "boot_types.h"
 #include "timer.h"
@@ -17,8 +16,8 @@ extern "C" {
 Boot_StatusTypeDef ComInit(void);
 Boot_StatusTypeDef ComTransmitByte(uint8_t data, uint32_t timeout_ms);
 Boot_StatusTypeDef ComReceiveByte(uint8_t *data, uint32_t timeout_ms);
-Boot_StatusTypeDef ComTransmitPacket(uint16_t msg_id, uint8_t *data, size_t length);
-Boot_StatusTypeDef ComReceivePacket(uint16_t *msg_id, uint8_t *data, size_t *length, uint32_t timeout_ms);
+Boot_StatusTypeDef ComTransmitPacket(uint16_t msg_id, uint8_t *data, uint8_t length);
+Boot_StatusTypeDef ComReceivePacket(uint16_t *msg_id, uint8_t *data, uint8_t *length, uint32_t timeout_ms);
 Boot_StatusTypeDef ComAck();
 Boot_StatusTypeDef ComNack();
